Moved simplify method
The simplify method in the Fraction class was being triggered at every call of the toString method, which is useless if it's a whole number. So I moved it to be within the conditional block that only runs if there is a fractional component to the object. Much more efficient, since the simplify method itself can be quite a drain.