Updated Android Studio
Upgraded to Android Studio 3.2. The automated process deleted several (apparently) now-useless files, and also deleted the code style guide and inspection profiles. After ensuring it was just a minor glitch, and there was no change in how the code style guide or inspection profiles were stored, I restored them. The result is only the useless files which I also did not write or care about, are the only ones deleted.