Updated build tools and SDK
Now using the latest build tools, and compiling for SDK 28. Had to update the support libraries as well, to version 28.0.0; but for some reason the support-v4 and support-media-compat libraries, which are not specified in the gradle scripts but I'm guessing are dependencies of other libraries, are being loaded as version 26.1.0. This is being identified by Android Studio as an error, but it is not causing problems with the build. So I'm thinking it's fine.