Fixed sudden errors
I'm guessing a change caused by the upgrade to SDK 28, likely from one of the support libraries, caused a few errors to pop up. The menu was misidentified, and there was a line in the ReferenceCell class that specified an int literal, rather than the constant it was supposed to use. I'm guessing it was fine before now because the use of the name 'menu' was a general term and there was only one menu so it was fine. Using the literal opposed to the constant was fine because the support library didn't specify it needed to be a constant. Though it's weird I didn't notice that error, since it's been like that for over a year (780da46a7a2f731e5008e2aaf1c5dd756b5bf51e). Anyway, both problems have now been fixed.