Fixed dependency error
The support library conflict caused by upgrading to SDK 28, as elaborated on in an earlier commit message (0026a7a80d431c03873746cd720cd31bf9f9b39a), has finally been resolved. I was forced to fix it because gradle was refusing to run a proper build as long as there were lint errors. The solution was simple: after a bit of research, I found a command that would allow me to properly examine which dependencies were being included, with a full tree structure and everything (./gradlew app:dependencies --configuration releaseCompileClasspath). This allowed me to trace the root cause of the problem to the firebase-core library. The same article specified multiple solutions, and I figured the simplest was implementing a resolution strategy, which should only run as long as the problem support library is included. So, I did that. As a result, everything should work now.