Added 'final' modifiers to QuestionManagement properties
Added 'final' modifiers to each of the properties in the QuestionManagement class. This is mainly because I can, and because it'll prevent the properties from being accidentally overwritten as development continues. Or at least it would, if the 'final' modifier actually worked on array elements, and not just the array itself. Oh well. It's the best I can do for now.