Moved most of the badges
The top of the readme was getting ridiculously crowded and unreadable because of all the badges from the various CI and coverage services. I reduced the number of badges at the top so there's only one in each category, and moved them all to a different part of the readme. Should look a lot better.