Added line graph
Added a simple line graph to the LogView. The biggest issue I had here was simplifying the X axis, which stores the date, so I don't have overlapping numbers. Because originally, even if I passed a standard java.util.Date object (which it does accept natively), the numbers on the X axis were in the millions, and the graph labels would overlap in that case. I thought of using the LogTypeConversion methods, but that would have problems of consistency, as the gap between the last day of one month and the first day of the next would be huge compared to any other day pairs. Eventually, by searching through the joda-time docs, I found a simple way to just calculate the number of days since a set epoch, which would be reliable and consistent, which I like, so I went that way. Beyond the issues with the date, I originally planned to have the graph separated by making the LogView a tabbed screen. But now I think having the graph at the top of the list of DailyLogItems is much better visually. Especially since the tabs themselves will take up some valuable real estate.