Added Firebase testing to Circle-CI config
The Circle-CI builds were only running unit tests, which is only half of the testing suite. So I wanted to add instrument testing, but found out the standard Android emulator is not supported on Circle-CI. So I configured the instrument tests to run through the Firebase Test Lab, which is a much more powerful solution anyway.