Fixed Circle-CI config... again
It appears the tutorial on the Circle-CI site was slightly flawed. But I think this will work. I had to manually set the locations of the app and testing apks, which I can only assume is the same as it is when I build locally, and remove the results-bucket option, as well as the fetching of the results. Because apparently, saving the results to a prescribed storage location is a paid-for service, and I refuse to spend any more money on this project than I already have.