Refactored for AndroidX
The Android support libraries are being replaced by the AndroidX libraries, which are several times better in my view, so I decided to refactor the project to take advantage of the new system. The process was completely automated by Android Studio. I would've done it sooner, but the Room libraries in AndroidX were still release candidates, until today. I also took the opportunity to revert an earlier commit (b703e6716ce2466c80d75a3686d9d7a2ed031ca3) that was intended to fix a dependency error that was causing an old version of a support library to be loaded, which was causing a conflict with the current support libraries, since all support libraries are required to be of the same version. But with AndroidX, this is not an issue. One can mix and match various versions of the AndroidX libraries freely without any problems. Though I'm pretty sure it's a good idea to keep them all somewhat up-to-date regardless.