Optimized the gradle build file
Modified the build file to only include Firebase and Crashlytics on the release build. Did the same to the joda-convert package as well, because it's only needed to fix some proguard issues, which is only used during the release build. Finally, I simplified the line including the espresso library, because I had no idea what the 'exclude group' line was for, and from what I can tell, if it was needed at some point it's not anymore, because new Android Studio projects don't include it.