Moved service badges to tables
To organize the analysis, CI, and coverage badges, I decided to place them all into tables. As part of that, I decided to take the opportunity to write short reviews of each service. Because that's really the reason I've been running so many of them, so I can compare them, and learn from them.