Modified log graph
Altered the graph in LogView in several ways. First, I moved the graph to be outside the ScrollView, so one can view it as they scroll through the individual log items. Also stylized the visuals of the graph, keeping the Y-axis between 0 and 100, hiding the X-axis labels, and making it horizontally scrollable and zoomable. Also had to specify the bound of both the highest and lowest dates, so the entire graph is visible, because for some reason, without doing that, it ends up being ridiculously zoomed in. To prevent any conflicts or crashes, I ensured that any edits to the UI are done on the UI thread. Which I didn't have to do, because it still worked even when I did all the coding in the AsyncTask, but I thought it was better to ensure that everything was in the proper threads, both for speed and stability.