Fixed no logs crash
The app was crashing whenever I tried to open the LogView on my phone, which I quickly realized was only happening because I didn't create any logs for the debug builds on that device. Eventually I traced the cause of the error. The changes made in a recent commit (b5389e5f8675b7d4e2c988c8c9c6da3bc35027fb), specifically the attempt to record the first date in the records, to place as a zero on the graph, was causing an ArrayIndexOutOfBoundsException on the empty array. I figured the simplest solution was the best. Just check if there are no logs, and return 0, bailing out of the AsyncTask.