Redesigned LogView for horizontal view
The LogView screen is now appropriately laid out when horizontal. Rather than a chart filling the whole screen, and a tiny letterbox view of the log item list; the chart instead only covers half the screen, with a scrollable list of log items on the right. It's a much nicer look.