Replaced Joda-Time with ThreeTen
In my investigations of open-source analytics, I ran across something quite interesting: A backport of the java.time package called ThreeTen. What this is, is the complete java.time package, wrapped up so one can import it into their Java 7 project. And because this app, despite using Java 8, can only use the Java 7 libraries, this was perfect. For a long time I was really annoyed that I couldn't use the official java.time package like I wanted to originally (897508252e5c3e3840eec9952d1b2f98875a14b8). So this feels like a relief. It has a much better design, in many ways. For instance, getting the date to display the day of the week as a word doesn't require an obscure formatting pattern like "EEEE". As a result, the code is now much more intuitive. And, on the off-chance I decide to just abandon all users running anything older than Oreo, I just need to change all the imports. It's the perfect drop-in replacement. So I'm very happy with this change.