Added AndroidThreeTen initialization
Turns out the AndroidThreeTen library, which I'm using instead of the standard ThreeTen backport for reasons of efficiency, needs to be initialized with the application context, otherwise any attempts to access time zone information will cause a crash. Which is weird, because this app doesn't use time zones. But it probably still needs the time zone database for fetching the current date. It's very unusual, because when testing on the Android Emulator, I had no issues. But when I tested it on my actual phone, only then did I get a crash. Anyway, this change should prevent it from happening at all in the future.