Altered preservation of KanaSelectionPage
Altered the code used to prevent the KanaSelection screen from clearing whenever the screen is reloaded (de136237ff16f369bfad0f5546a55e521f1e130d), by having the saving code run a bit earlier in the lifecycle chain. This is because the code was not triggering when the app was cleared out by Android's built-in memory-saving features. This isn't the first time this feature has caused me problems (6e4372ef07ca7addce59f85cb7fb20256835e30c). Mainly because, I'm finding it hard to test for that scenario.