Created 'pure' build type for F-Droid
Created a new 'pure' build type with Firebase and Crashlytics removed, and therefore uses only open source libraries. This is mainly for the benefit of the F-Droid repository which only accepts pure open source projects with no closed source libraries. This 'pure' build type instead uses ACRA crash reporting, which I set up to use a special email address that I set up exclusively for this. This process of creating this build type was quite annoying. Because ACRA, unlike Firebase and Crashlytics, requires a special initialization routine. So, I had to find a way to get it to use an alternate Application class that extended the original Application class and runs the initialization routines. But all my attempts to do so failed. I tried to integrate the initialization routines into the KanaQuiz class as it was, but because I'm not including the ACRA libraries in all builds, that failed at the compilation stage. The only work around I found was to rename the KanaQuiz class, mark it abstract, have dud classes in the debug and release builds that did nothing but extend the original class, and place all the initialization code in the 'pure' build's KanaQuiz class. I'm not sure I like this, it's not the most elegant solution, with those pointless dud classes that only exist to rename the Application class. But it does work, without producing redundant code. So I guess I'll have to deal with it.