Disabled Proguard on pure build
Proguard was causing problems on the pure build by looking for Firebase stuff that are not supposed to be included in that build. I tracked the source of the issue to the Firebase performance monitoring gradle plugin, likely injecting code into the build. But because I have not found a way to disable any plugins without disabling them on all builds, I decided to instead disable Proguard. This is intended to be a temporary solution.