Implemented vocabulary questions
Did a lot of work implementing the vocabulary questions feature. It's not quite complete. Still need to implement a few auxiliary features, such as renaming a few variables, methods and classes to fit the fact that it's not all kana anymore; and having the user select how they want to view the vocabulary questions. But all the important stuff works, including the selection screen and reference. So I'm happy. The questions are also preliminary. Hoping to add quite a few more blocks in the near future.