Disabled Vocabulary Reference
Been having problems with the Reference screen. In a three tab layout, with tabs contained within, if one goes from the first tab, to the last tab, and back again, the contents of the subtabs will vanish. If one navigates through the subtabs, the contents could come back if there are enough tabs, but they shouldn't disappear in the first place. I've narrowed the problem down to a simple issue. When tabbing through the main tab set, the pages within are being cleared out of memory, through the standard procedure. However the pages within those pages are being cleared out of memory without going through the procedure, and as a result, are not being restored properly when they return. I know this because I checked and none of the various onStop/onStart/onPause/onResume/onDestroyView/onCreateView methods for ReferenceSubsectionPage objects are being triggered at all during the top level tab transitions, despite the fact that they are triggered in the Fragments that contain these objects. I spent all of yesterday trying to fix this, and have officially given up. The only solution available to me now is to just disable the Vocabulary tab until I fix it. So that's what I'm doing... for now.