Added kanji to vocabulary file
I looked up the kanji for various words and added them to the vocabulary questions file. Some words, it turns out, don't have kanji; or, in the case of 'iie' (no), there is a kanji that one could use, but it's not normally used, so I didn't include it in the definition, but noted it in a comment. I also found an interesting note about the word 'anata,' which is it can be gendered, but it doesn't affect the pronunciation, just the kanji representation. I added a comment explaining that as well, just in case I decide to do something about it later. I also replaced the hiragana used for 'noto' with the katakana, which apparently is the proper character set to use. I also modified the schema definition file by slightly expanding the kana type definition to include the entire set of unicode characters in the hiragana and katakana blocks. This is mainly because I needed to use the prolonged sound mark in the katakana for 'noto,' which wasn't included, but also because I don't want to come back here and add another character every time this comes up. All I added to the definition is a few obsolete and unused kana, punctuation and iteration marks, something labeled 'koto,' individual diacritics, and something labeled 'yori.' This is still all technically kana, and I didn't include the unused or 'reserved' codepoints. So it's narrow enough to still keep it's purpose, but broad enough to include everything that might come up. It should work out well.