Redesigned KanaQuestion classes
New Question class will be used as a generic abstract version of the KanaQuestion class, extended by the KanaQuestion and WordQuestion classes. I redesigned many classes so they use the Question class, instead of the KanaQuestion class. This makes it easier to predict certain app behaviour. Since the WordQuestion class needs to handle many things differently, it's simpler to just extend a generic class, rather than risk the KanaQuestion methods accidentally being triggered because I didn't override the methods properly. Also, the methods have been renamed to be no longer KanaQuestion specific, and therefore their function is clearer. I'm not quite done with this process. I still have to rename some things in the ReferenceCell class, and the 'options' and 'logs' packages. But for now, this is good.