Edited properties and methods in logs package
The properties and methods in the various log classes have been renamed to no longer be kana-specific, changing 'kana' and 'romanji' to 'question' and 'answer'. I even renamed the KanaRecord class to QuestionRecord. I didn't change any database names because that would require writing a migration script, and that seems a bit overkill at this stage. But I did make a note reminding me to make that change the next time I decide to alter the database schema.