Added to upcoming feature list
Added some new feature ideas to the upcoming feature list. The ideas I have are to enhance the KanaSelection screen to allow users to select individual kana, and not just whole groups; and to allow the user to make their own question lists, which'll likely be stored in the database.