Added update alert plugin
Added the Gradle Versions Plugin to check for available updates to library dependencies. I've been using it for a while, but just decided not to add it to version control because it was not needed for any builds. But because I was sick of having to add and remove it whenever I wanted to properly check for updates, I thought it was time I added it properly. It won't damage anything by keeping it in.