Altered Crashlytics declaration
Removed the 'transitive' property from the declaration of the Crashlytics dependency. This is because I didn't know exactly what it meant; it was interfering in Android Studio's built-in checks; and after a quick look at the Crashlytics documentation I noticed the 'transitive = true' line has been removed along with the 'aar' specification, meaning they're no longer necessary. So that's enough reason to remove both of them, I feel. Much simpler.