Redesigned utility classes
I was using the abstract modifier for utility classes, which are classes containing nothing but static methods. But I since learned that this was probably not best-practice, and it was better to set them as final classes with private constructors. That way the classes can't be extended or instantiated. So that's exactly what I did.