Altered how displayContents method for WordQuestions
The KanaSelectionItem objects for sets of WordQuestions have been displaying the actual questions in their romanji forms (or kana/kanji forms, depending on the user's preferences). This isn't really the most readable form they could be in. So I changed it to display the English translations instead, comma-separated. As part of this, I had to add a bit of code to remove the last comma, if it's there, because otherwise it looks confusing.