Cleaned up Gradle scripts
The two build.gradle files have been cleaned up a bit. I moved the android-junit-jacoco plugin to the plugins block in the project-level build script, and removed the Google repository from the app build script, since it's already in the project-level one. I also replaced the use of the custom declaration, specifying the URL of the Google Maven repository, with a simpler declaration of the same repository, now using the code built into Gradle. It's much nicer, and overall more consistent. I like it.