Disabled closed-source plugins on pure build
I finally figured out a reliable way to stop the Firebase plugins from running on the pure build. It's a simple 'if' statement that will only be true if it's running a task with the word 'release' in it, which should only be true if we're running a release build. I tested it, and it seems to work. I don't get the build errors I once did while trying to run a pure build, and the Firebase stuff on the release build still works, so I'm satisfied. I also reenabled Proguard on the pure build, since, as I said, it stopped throwing the errors I saw before (64af17954bd5527565cd85a2548f8c661b02a3e8). This should be enough to satisfy any Stallman-esque open-source pedants out there.