Renamed toString method
The toString method in the Question class was intended to be used for looking up entries in the database, since the WordQuestion objects could display the question in several different forms, toString would always display it in romanji so the database wouldn't have multiple entries for the same question storing different data. And the database calls would always look up the same entry for the same question. But I realized that could be a bit obtuse. So I renamed the toString method to the getDatabaseKey method. That's much clearer.