Fixed tab transition issue
The issue I was having with the reference screen, with pages within pages not being cleared properly, and therefore not being rebuilt properly (as explained here: 4099cfe709524fd0c1834a69a9421e8543ee08dd), has finally been fixed. The solution was ridiculously simple. I just had to replace the activity's FragmentManager, with the FragmentManager for the pages containing the other pages. So I made the change, and de-commented the code for the vocabulary reference tab. Very happy with this result.