Added no reference hook
Added a hook to the reference screen for it to display a special message if no reference items are to be displayed. This should only happen if no items are selected on the KanaSelection screen, and the full reference option is disabled, which would be an odd scenario. But without this hook, we just have a blank screen in this scenario, which tells the user nothing, and could mean the page is still loading. So I thought, we should display a message here.