Fixed Reference Screen
The check I built for an empty reference screen (4fce65ff7ed1237b44f27ba7ba3df45e6b660f26) wasn't done properly. It's my fault, I didn't properly test it. I only tested it in the empty scenario, I didn't think to test it for the non-empty scenario at all. I just assumed it would work, but it was actually triggering all the time, and clearing out the reference screen even when it wasn't supposed to. Fixed it by having it check the count of the ReferencePager, instead of the child count of the ViewPager.