Added another feature idea
Added a feature idea to the Upcoming Features section of the readme. It's an idea for an option to add Furigana (a special reading-aid that adds the Hiragana version of a word just above the Kanji) to the vocabulary questions. Will probably require adding an additional attribute to the WordQuestion xml tags, since Katakana text can have Furigana, but they would have to be in the Hiragana format. Might also require having two separate options, one for Furigana on Kanji text, and the other for Furigana on Katakana text. Also worth remembering the <ruby> tag, which is an HTML tag that can be used for displaying Furigana, since here are ways to get TextView objects to display HTML-formatted text, but it might require some jiggering to get it to work.