Expanded isAcraInstalled method
The isAcraInstalled method now checks if ACRA has been properly initialized, because before it only checked if the libraries were installed. I had to do this with reflection, because the required libraries won't always be available at compile time. As part of this, I had to modify the try-catch to catch many more exceptions. I tried to merge it into a single catch statement, but apparently it'll compile into a ReflectiveOperationException, which only exists in Android 4.4 KitKat (API 19). So I had to split it up, ignore all the exceptions, and have it fall into the final 'return false' statement. Simplest solution really.