Added new feature idea
Added an idea to the upcoming features list. The idea is to give users the option to export logs and preferences to an external file, likely in XML or JSON format. This is mainly inspired by the fact that FDroid packages, which is a repository this app will soon be available on, will be signed with their own key, which is not the same as the one I use on Google Play. So anyone migrating between those two versions will have to uninstall the old version first, and lose all logs and preferences. The only solution I can think of to this issue is allowing the user to export their stuff to an external file, which'll be quite annoying no matter how well I program it. Though it won't be the least user-friendly thing about that build, if we're being honest.