Removed question prefIds from string resources
Because the prefIds were only used in the question XML documents, and didn't need to be translated in any way, I decided to remove them from the list of string resources, and have them placed as literals in the question XML documents. Don't know if it's more efficient, but it just makes more sense to me. The only issue that popped up was the fact that the OptionsControl class did use the string resources to define the one question set that's enabled by default. Easy fix for it to instead just pull the same information from the parsed XML document in the QuestionManagement class.