Modified romanji type definition
Edited the romanji type in the xsd file to remove a few ancillary characters that I accidentally included. They exist in the unicode table between 'Z' and 'a', but shouldn't exist in this definition. Also added several characters that I've learned can be included in romanji; that being vowels with macrons over them (lines that define them as long vowels), and the apostrophe and hyphen. This is a much more accurate definition.