Added numbers to vocabulary
Added as many numbers as I could to the vocabulary questions. Didn't include any that are formed by combining other numbers, because I thought that could needlessly complicate things, and I'd like to keep it simple. Besides, if you know 'go' is 5, and 'ju' is ten, then you should know 'ju-go' is 15, and 'go-ju' is 50. I might decide to add combinations in future.