Added people category in vocabulary
Added some vocabulary questions for people and family. Also altered the schema definition allowing kanji datatypes to be surrounded by hiragana and still be valid. This is mainly because for certain words, and it appears to be more formal words, the kanji itself is not enough, and it needs to be surrounded by hiragana. The specific kanji can stand on their own, but they have completely different pronunciations in those cases, and I want the questions to be about specific pronunciations. So this change is quite vital, but it might interfere with my plans of adding a furigana feature later on.