Fixed vocabulary reference tabs
There are a lot of vocabulary categories, and because I've never had a tab list of this size before, I wasn't prepared for what effect that would have. Turns out, as the list grows, the tabs get smaller, and the names have to be truncated. But I fixed this by making the tab list scrollable. The user can flick the tabs to the left or to the right, and see more tabs. Unfortunately there doesn't seem to be a reliable way to check if it is necessary to make the list scrollable, and if it isn't, the tab list looks weird with them all huddled together on the left. So I had to include a specific hook so the only scrollable tab lists are those with more than four tabs. The only issue with this method is that if the screen is of a size smaller than the minimum screen size I'm expecting, it'll still truncate the tab names. But I can't see a more reliable solution to this. So it'll have to do. I could probably do some complected coding to solve this, involving measuring the screen, and the tabs, and stuff like that. But it's a bit too much effort than I think is warranted at the moment. So this'll have to do.