Added food and drink category in vocab
Added several vocab questions on the subject of food and drink. Also, continuing a recent change to the schema file (71b9086e19104973007c70c90190fc7a37391c6b), I adjusted the kanji datatype definition to basically allow any combination of hiragana and kanji, in any order, as long as there's at least one kanji character within. It's not the most elegant in terms of execution, given the repeating sections of the pattern, but since when are regular expressions elegant?