Rearranged a few vocab categories
I decided to change the order of a few of the vocab question categories. It may seem like an arbitrary decision, and it is, but I decided to do it because I wanted it to match the question order in Learn Japanese to Survive: Hiragana Battle, which, full disclosure, is what I'm basing much of this app on. I thought if the makers of that game ordered the questions in that way, there might be some logic to it, and I might as well follow it because there's no logic to what I'm doing, other than that's the order I found my notes in.