Fixed QuestionSelectionItem attribute file
When I renamed KanaSelectionItem to QuestionSelectionItem (2791d9b6474e60c73ec37106df51885078c53060), I forgot to rename the file defining the XML attributes, and connect it to the renamed class. I didn't notice because I don't actually use the XML attributes. But nonetheless, I thought it was a good idea to update it.