Set unique restriction on primary keys
The various question objects in the xml question documents must be unique. Having the same question twice in the system would be a very bad idea. Mainly because the database references the questions, and having the same question pop up twice would mean the same database entry would be referenced twice, causing a doubling-up of information. The questions should reference separate entries, and separate primary keys. This took quite a bit of work, since most of the sites I referenced seemed to have incomplete information. While I got the syntax down pretty quickly, for some reason the selectors had to be referenced by specifying namespaces, which I find a bit odd, since the current target namespace should be the default. So I had to specify a specific namespace for this document. Then there was the issue of location. Turns out the scope of the unique check is specified by its location, and placing the check in the element it specifies is useless, it must be placed in a parent object. In fact the root object is best, especially for this purpose, so it checks the whole document. But then you have to specify how deep in the hierarchy the elements being checked are. You can't just specify it to check any item at any depth. So I had to have the KanaQuestion selector specify multiple depths, because they can exist at multiple depths. Anyway, it now works. I could specify that there's no duplication across the two types, but it's pointless, since the primary keys have completely different, and non-overlapping restrictions. Also, the unique check cannot work across multiple documents anyway. Besides, even if I wanted to, and it was necessary, I don't know how to do that. While I can specify multiple selectors, the fact that the primary keys have different field names will be a real issue.