Removed macrons from romanji
I realized that the macrons, while valid as romanji, aren't really necessary. Because of this, any time a romanji form can contain macrons, it can also not, which means there are two possible versions of the question. For the sake of consistency, I decided to remove macrons from the romanji datatype definition, and rewrite all items that use macrons, to go without. As part of my research on this, I also looked up all the possible romanji conversions for every question, to find any that could use macrons. For future reference, I recorded that information in a comment to the right of each appropriate question. Will probably never use it, but it's pointless to let that work go to waste.