Changed default for vocabulary question format
Thought it would be better for the default format for vocabulary questions to be kana, and not romanji. This is, after all, The Kana Quiz. One can still change it to romanji, if they wish. Also edited the questions schema document to set the kana field as required in WordQuestion items. Not for any technical reason, just because there is no reason to not include it in the WordQuestion items.