Added a few more unique restrictions to xml schema
Added a few additional unique item restrictions to the question schema definition file. One specifying that each prefId can only be used once, and another specifying that no two sections can have the same definition within each KanaSet. Also added a note saying I need to find a way to disallow a no diacritics and no digraphs section, because that's the default for those questions outside a section, and it would cause issues in the parser if a section was defined with those parameters.