Added more vocab categories
Added three additional vocabulary categories, which I sourced from Learn Japanese to Survive: Katakana War. There is a category overlap. Yet another with household items. So I could merge the categories, but let's be honest, that would be a bit nuts. Having a single category with nearly 20 items is overkill, and would make it difficult for progression. I think I might want to cap the number of questions per category at ten or 12. So having a second House & Home category is perfectly reasonable.