Removed big numbers
Big numbers are a big problem. For a start, once one gets to the 10,000 mark, you have to always add a number to the front of it. It doesn't just default to one, or 'ichi'. So having just 'man' exist as a question is just incorrect. It should be ichiman. But if I include that, I should also include niman, and sanman... and on and on. So I might have to deal with this a different way. Also, asking the user to parse what ten quadrillion is, is a bit of an ask. Also, I did consider expanding the list, and found out the number 'gougasha' is equal to 10^52, or 10^56. There is no definitive answer I can find on the subject, which tells me there is minutiae I'm not qualified to elaborate on. So I decided to just cut it down. I think I might have to include this as a special feature at a later date, where numbers are constructed in a very special way. But for now, I'm just removing it.