Added a couple feature ideas
Added some additional feature ideas to the upcoming features list. A pronunciation guide would just be audios of people pronouncing the individual items. But because this is a non-profit, solo-developed project, unless I find a native speaker willing to volunteer their voice, or I can find a royalty-free library, it will probably be me recorded through a mediocre microphone. That might be a bad idea, not sure at this stage. Also, number questions would likely involve a whole new extension to the QuestionManagement class, which'll involve custom settings of combinations of glyphs and length limits binding together all possible numbers, so I won't have to manually enter every number from 0-10,000 into an xml document. Finally a reverse question mode, which'll just reverse how the questions currently work. Instead of asking what the romanji equivalent of a character is, the question will be which character belongs to this romanji. I imagine this'll require some complex coding.