Added note reminding me to fix the tab size issue
The issue with tab sizes I mentioned before (c2eb24fb414b87ede6b4ecb66fa9620eb5132a10), is more of an issue than I thought, since I didn't think to check for when the screen is horizontal. But I'm not sure exactly how to fix it right now, so I just added a note to remind me to fix it at a later date.