Adjusted ReferenceCell construction
I was a bit annoyed with the vocabulary reference screen, because it was displaying characters at inconsistent sizes. If the item was only one character, either a single kana, or a kanji, it would display it at full size; but if there was more than one character, it would be slightly smaller. But that's only because I was using the same code that I used for the KanaQuestions, unchanged. So, I modified the code so that ReferenceCells for WordQuestions always use the slightly smaller font, and ReferenceCells for the KanaQuestions are unaffected. This'll look much nicer.