Normalizing multiple choice buttons
The sizes of the multiple choice buttons were really inconsistent when doing vocabulary questions, which didn't look very good. So I decided I needed a fix. Writing the code for this was initially very simple. Just find the largest button of the set, and change the width of all buttons to match. But the problem quickly came up that it wouldn't trigger properly. Or to be more precise, it would trigger, but instead of setting all buttons to the largest size, it set them to the minimum possible size the class would allow. I realized very quickly why this was happening. Until the button objects have finished being built and placed within the Activity, the getWidth method will always return a size of zero, because the sizes of the buttons themselves have not been properly calculated yet. So I tried setting a 100ms delay on running the code, and that worked, except when the activity was first loaded, because the activity was always waiting until everything was ready before it would assemble the view. Eventually, I found a solution. Use an AsyncTask to multithread the operation. So the second thread would check if its attempts to find the largest button were meaningful, over and over again, with a 100ms delay between attempts, until it got a proper result, without locking up the UI thread. So only when the UI thread was ready would it actually get a proper result, and use that to resize all the buttons. I also made the buttons go completely invisible while this process is running, so we don't have the buttons resizing before the user's eyes. Since that would look quite odd. I might have to readjust certain elements of this, but for now, it works.