Reusing StringBuilder
The StringBuilder object I was using for the parseRomanizationSystemList method in XmlParser used a StringBuilder object to properly segregate the elements of the String. But when it was time to build a new item, it was creating a new StringBuilder object, which left the old StringBuilder object to wait around for the garbage collector. Much more efficient to just reuse the StringBuilder object instead. So, by setting the length to zero, I emptied it out for reuse.