Simplified code for parseRomanizationSystemList
Can't believe I didn't think to look for this before, but I found a method in the String class that would automatically split up a string into an array of strings, with a pre-defined separator. And with that I was able to dramatically simplify the parseRomanizationSystemList method. No need for List objects, when I can use primitive arrays. And no need for StringBuilders when I don't have to loop through every single character. Much simpler.